MCH - Update that completes the Leaderboard Game Center integration by adding the screen invoke and by adding the call the record the high score.

+++ Invaders/MCHGameCenterManager.h
@@ -17,7 +17,7 @@
 @property (assign, readonly) BOOL gameCenterAvailable;
 
 + (MCHGameCenterManager *)sharedInstance;
-+ (void) reportScore: (Float64) score forIdentifier: (NSString*) identifier;
-- (void)authenticateLocalUser;
++ (void) reportScore: (int) score forIdentifier: (NSString*) identifier;
+- (void) authenticateLocalUser;
 
 @end
